simulator: Refactor out of crate feature
- Move `simulator` example into its own crate
with all its dependencies

- Remove crate feature `simulator`, code implementation
is moved to the `simulator` example crate

- Created workspace to host `mousefood` crate + examples + future stuff
