+++ Cargo.toml
@@ -1,3 +1,10 @@
+[workspace]
+resolver = "3"
+members = [
+  ".",
+  "examples/*",
+]
+
 [package]
 name = "mousefood"
 description = "embedded-graphics backend for Ratatui"
@@ -14,7 +21,6 @@ exclude=["/.github", "/assets"]
 [dependencies]
 ratatui = { version = "0.29", default-features = false, features = ["all-widgets"] }
 embedded-graphics = "0.8.1"
-embedded-graphics-simulator = { version = "0.7.0", optional = true }
 embedded-graphics-unicodefonts = { version = "0.1.0", optional = true }
 weact-studio-epd = { version = "0.1.2", features = ["blocking"], optional = true }
 
@@ -24,13 +30,8 @@ paste = "1.0.15"
 
 [features]
 default = ["fonts"]
-simulator = ["dep:embedded-graphics-simulator"]
 fonts = ["dep:embedded-graphics-unicodefonts"]
 epd-weact = ["dep:weact-studio-epd"]
-
-[[example]]
-name = "simulator"
-required-features = ["simulator"]
 
 [lints.rust]
 missing_docs = "warn"
