+++ examples/simulator/readme.md
@@ -1,0 +1,19 @@
+# Simulator
+
+Run mousefood apps on your computer inside a simulator! Uses [embedded-graphics-simulator](https://crates.io/crates/embedded-graphics-simulator).
+
+## Requirements
+
+This app requires [SDL2](https://wiki.libsdl.org/SDL2/Installation) to be installed.
+
+If you use [nix](https://nixos.org) you can run `nix-shell -p SDL2` before running the application.
+
+## Run
+
+To start this demo, simply run:
+
+```shell
+cargo run
+```
+
+A window will open with the simulator running.
